unisal_predict/pyproject.toml
[project]
name = "unisal-predict"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.3.3",
    "opencv-python>=4.11.0.86",
    "pillow>=11.3.0",
    "scipy>=1.16.2",
    "tensorboardx>=2.6.4",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
]
